Move PromptTemplateNode from examples into morpheus/llm
